fdinfo: Sanitize types in fdinfo_entry
The namelen is u16, to cover the PATH_MAX u8 is not enough.
The pos is u64, since file offset is that long indeed.
The id is u32 as per previous patch.

Fix printf-s respectively.

+++ include/image.h
@@ -46,11 +46,11 @@ enum fd_types {
 
 struct fdinfo_entry {
 	u8	type;
-	u8	len;
+	u16	len;
 	u16	flags;
-	u32	pos;
+	u64	pos;
 	u64	addr;
-	u64	id;
+	u32	id;
 	u8	name[0];
 } __packed;
